src/evolite/core/compression.py
from typing import Optional, Dict, Any
import numpy as np
from ..models.neural_net import CompressibleNetwork
from ..utils.helpers import calculate_compression_ratio


def compress_network(
    network: CompressibleNetwork,
    target_size: Optional[float] = None,
    compression_config: Dict[str, Any] = None,
) -> CompressibleNetwork:
    """
    Compress a neural network using evolutionary optimization.

    Args:
        network: The neural network to compress
        target_size: Target compression ratio (0.0 to 1.0)
        compression_config: Additional compression parameters

    Returns:
        CompressibleNetwork: Compressed network
    """
    raise NotImplementedError("Coming soon!")
